+++ laws/src/main/scala/spire/laws/AdditiveSemigroupLaws.scala
@@ -8,11 +8,14 @@ trait AdditiveSemigroupLaws[A] {
   def plusAssociative(x: A, y: A, z: A): IsEq[A] =
     S.plus(S.plus(x, y), z) <=> S.plus(x, S.plus(y, z))
 
-  def sum1(a: A): IsEq[A] =
+  def sumN1(a: A): IsEq[A] =
     S.sumN(a, 1) <=> a
 
-  def sum2(a: A): IsEq[A] =
+  def sumN2(a: A): IsEq[A] =
     S.sumN(a, 2) <=> S.plus(a, a)
+
+  def sumN3(a: A): IsEq[A] =
+    S.sumN(a, 3) <=> S.plus(S.plus(a, a), a)
 
   def trySum(xs: Vector[A]): IsEq[Option[A]] =
     S.trySum(xs) <=> xs.reduceOption(S.plus)
