+++ laws/src/main/scala/spire/laws/CRingLaws.scala
@@ -1,0 +1,12 @@
+package spire.laws
+
+import spire.algebra.CRing
+
+trait CRingLaws[A] extends CRigLaws[A] with CRngLaws[A] {
+  override implicit def S: CRing[A]
+}
+
+object CRingLaws {
+  def apply[A](implicit ev: CRing[A]): CRingLaws[A] =
+    new CRingLaws[A] { def S: CRing[A] = ev }
+}
