+++ laws/src/main/scala/spire/laws/MultiplicativeGroupLaws.scala
@@ -5,13 +5,13 @@ import spire.algebra.MultiplicativeGroup
 trait MultiplicativeGroupLaws[A] extends MultiplicativeMonoidLaws[A] {
   override implicit def S: MultiplicativeGroup[A]
 
-  def leftNegate(x: A): IsEq[A] =
+  def leftReciprocal(x: A): IsEq[A] =
     S.times(S.reciprocal(x), x) <=> S.one
 
-  def rightNegate(x: A): IsEq[A] =
+  def rightReciprocal(x: A): IsEq[A] =
     S.times(x, S.reciprocal(x)) <=> S.one
 
-  def consistentMinus(x: A, y: A): IsEq[A] =
+  def consistentDiv(x: A, y: A): IsEq[A] =
     S.div(x, y) <=> S.times(x, S.reciprocal(y))
 
 }
