+++ laws/src/main/scala/spire/laws/SemigroupLaws.scala
@@ -8,11 +8,14 @@ trait SemigroupLaws[A] {
   def combineAssociative(x: A, y: A, z: A): IsEq[A] =
     S.combine(S.combine(x, y), z) <=> S.combine(x, S.combine(y, z))
 
-  def combine1(a: A): IsEq[A] =
+  def combineN1(a: A): IsEq[A] =
     S.combineN(a, 1) <=> a
 
-  def combine2(a: A): IsEq[A] =
+  def combineN2(a: A): IsEq[A] =
     S.combineN(a, 2) <=> S.combine(a, a)
+
+  def combineN3(a: A): IsEq[A] =
+    S.combineN(a, 3) <=> S.combine(S.combine(a, a), a)
 
   def tryCombine(xs: Vector[A]): IsEq[Option[A]] =
     S.combineAllOption(xs) <=> xs.reduceOption(S.combine)
