+++ laws/src/main/scala/spire/laws/VectorSpaceLaws.scala
@@ -11,7 +11,7 @@ import org.scalacheck.Prop._
 
 object VectorSpaceLaws {
   def apply[V: Eq: Arbitrary, A: Eq: Arbitrary: Predicate] = new VectorSpaceLaws[V, A] {
-    val scalarLaws = RingLaws[A]
+    val scalarLaws = OldRingLaws[A]
     val vectorLaws = OldGroupLaws[V]
   }
 }
@@ -20,7 +20,7 @@ trait VectorSpaceLaws[V, A] extends Laws {
 
   implicit def scalar(implicit V: Module[V, A]): Rng[A] = V.scalar
 
-  val scalarLaws: RingLaws[A]
+  val scalarLaws: OldRingLaws[A]
   val vectorLaws: OldGroupLaws[V]
 
   import scalarLaws.{ Equ => EqA, Arb => ArA }
